cc 6

+++ source/dungeon/gendung.cpp
@@ -1059,7 +1059,7 @@ static void DRLG_FTVR(unsigned offset)
 {
 	BYTE *tvp = &dTransVal[0][0];
 	// do not crash if the tileset does not match
-	if (offset >= lengthof(dTransVal) * lengthof(dTransVal[0]))
+	if (offset >= (unsigned)lengthof(dTransVal) * lengthof(dTransVal[0]))
 		return;
 	if (tvp[offset] != 0) {
 		return;
